Added partyPreviewObj state and state setter, added Partial typing when creating Context
partyPreviewObj state allows using useEffect to update data in Preview component correctly, and cuts down on necessary logic. Deleted commented code. Using Partial typing for now but I hope to be able to remove this, as it makes the other types useless and is causing a number of null checks in components receiving the firebase data