Fixed typing in Context to remove need for Partial typing. Updated Dropdown and Preview logic
Removed optional chaining and non-null assertions in Dropdown and Preview now that there is no issue corresponding to states potentially being undefined. Updated BaseDataType, PartyPreviewDataType, FirebaseDataType in Context as well as init states for all of them so that Partial typing is no longer needed for the useContext default state.